+++ src/DotNet/Writer/PESection.cs
@@ -69,13 +69,13 @@ namespace dnlib.DotNet.Writer {
 
 			writer.WriteBytes(Encoding.UTF8.GetBytes(Name + "\0\0\0\0\0\0\0\0"), 0, 8);
 			writer.WriteUInt32(vs);			// VirtualSize
-			writer.WriteUInt32((uint)rva);	// VirtualAddress
-			writer.WriteUInt32(rawSize);		// SizeOfRawData
+			writer.WriteUInt32(rva);		// VirtualAddress
+			writer.WriteUInt32(rawSize);	// SizeOfRawData
 			writer.WriteUInt32(dataOffset);	// PointerToRawData
 			writer.WriteInt32(0);			// PointerToRelocations
 			writer.WriteInt32(0);			// PointerToLinenumbers
-			writer.WriteUInt16((ushort)0);	// NumberOfRelocations
-			writer.WriteUInt16((ushort)0);	// NumberOfLinenumbers
+			writer.WriteUInt16(0);			// NumberOfRelocations
+			writer.WriteUInt16(0);			// NumberOfLinenumbers
 			writer.WriteUInt32(Characteristics);
 
 			return alignedVs;
